Better resume after crashes, and option to download the data for just one user
A) Previously, you were required to deleted the old output folder before you ran the script.
You no longer need to do that. Now, if a target file already exists, then it will not
re-download it. This should be an effecient way to continue after a random crash.

If you really do want a clean re-download of everything, then delete the output file as before.)

B) There is a new command-line option, so you can do

node . --download-only X1234567 instruction-file.txt

(And in this case, it will skip the 'file already exists' logic, and really download everything.

C) Finally, the time-out for saving a PDF was increased to 5 minutes.